==> code/5.merge-results.ipynb <==
'''
This scripts merges the results of the classification
done manually with the original data collection.
'''

import glob
import json
from os.path import basename
import pandas as pd

def load_classification_data(path: str):
    """
    Load JSON files from the specified path and return a dictionary containing the data.

    Args:
    - path (str): The path where the JSON files are located.

    Returns:
    - A pandas DataFrame containing the loaded JSON data, with row numbers as the index and data as columns.

    Example:
    >>> data = load_classification_data("../../output/mvp/classification-2023-08-23/")
    
        classification
    0   fp
    1   d
    2   nd
    3   nd
    4   d
     

    DISCLAIMER: The documentation for this function was written by ChatGPT.
    """
    # Initialize an empty dictionary to store the loaded data
    data = {}

    # Get a list of all JSON files in the specified path
    files = glob.glob(path + "/*.json")

    # Iterate over each file
    for file in files:
        # Extract the row number from the file name
        row_no = int(file.split("/")[-1][:-5])

        # Open the file and load the JSON data
        with open(file, "r") as f:
            d = json.load(f)

        # Add the loaded data to the dictionary with the row number as the key
        data[row_no] = d

    return pd.DataFrame.from_dict(data, orient='index')


def main():
    # Read CSV
    df = pd.read_csv("../../output/mvp/3.tweets_with_keyword_detection/tweets-with-keywords.csv", sep="|")
    
    # Save a subset of the df without the keywords
    df_no_keywords = df[~df.has_both_keywords]
    
    # Keeps only those with the keywords
    df_keywords = df[df.has_both_keywords]

    # Loads the classification data as a separate df
    classification_df = load_classification_data("../../output/mvp/4.manual_classification//")
    
    # Joins the classification
    df_keywords = df_keywords.join(classification_df)
    
    # Saves
    df_no_keywords.to_csv("../../output/mvp/5.merged_dataframes/no-keywords.csv", index=False, sep="|")
    df_keywords.to_csv("../../output/mvp/5.merged_dataframes/keywords.csv", index=False, sep="|")


if __name__ == "__main__":
    df = main()
