==> code/8.prepare-dfs-for-sentiment-analysis.ipynb <==
'''
This script takes the files with the replies
to the high engagement tweets about environmental
protests, which were saved in JSON files, and
puts them into a dataframe.
'''

import glob
import json
import pandas as pd

def main():
    

    #####################
    ### Top 1 percent ###
    #####################
    
    # Reads tweet files
    fpaths = glob.glob("../../output/mvp/6.top_n_replies/*.json")
    
    # Adds tweets to a list
    dataset = []
    for fpath in fpaths:

        with open(fpath, 'r') as f:
            data = json.load(f)
            dataset.extend(data)
            
    # Turns into dataframe
    df = pd.DataFrame(dataset)
    
    # Saves results
    df.to_csv("../../output/mvp/8.dfs_for_sentiment_analisys/protest_tweets/protest-tweets.csv", sep="|")
    
    ############################
    ### Random sample of 100 ###
    ############################
    
    # Reads tweet files
    fpaths = glob.glob("../../output/mvp/7.random_sample/*.json")
    
    # Adds tweets to a list
    dataset = []
    for fpath in fpaths:

        with open(fpath, 'r') as f:
            data = json.load(f)
            dataset.extend(data)
            
    # Turns into dataframe
    df = pd.DataFrame(dataset)
    
    # Saves results
    df.to_csv("../../output/mvp/8.dfs-for-sentiment-analysis/random_sample/random-sample.csv", sep="|")

if __name__ == "__main__":
    main()
