==> code/9.high_engagement_language_detect.ipynb <==
!pip install transformers datasets

!pip install tensorflow

from datetime import timedelta
from google.colab import drive, runtime 
import numpy as np
from os.path import isfile
import pandas as pd
from transformers import pipeline
import time

# Mounts the Google Drive so we can access the CSV file
drive.mount('/content/gdrive')

# Reads the CSV file with the classified tweets
df = pd.read_csv('gdrive/My Drive/DW - Climate Protests/8.dfs_for_sentiment_analysis/protest_tweets/protest-tweets.csv', sep="|",
                 dtype={
                     "tweet_id": str,
                     "conversation_id": str,
                     "in_reply_to": str,
                 })

# Making sure that all the dtypes are fine
df.tweet_id.isna().value_counts()

# How many entries do we have?
df.shape

# Loads a pre-trained language detector so we can select only the tweets that are in English 
classifier = pipeline(task="sentiment-analysis", model="papluca/xlm-roberta-base-language-detection")

# Does it recognize English?
classifier("This is a test")

# What about Portuguese?
classifier("Isso é um teste")

# What about arrays?
classifier(["Isso é um teste", "This is a test"])

# Runs it through all tweets and saves the answer back to an array
def lang_detect(data, classifier):

  result = classifier(data.raw_content.tolist())
  result = pd.DataFrame(result)
  result = result.rename(columns={"label":"lang_label", "score":"lang_score"})

  return pd.concat([data.reset_index(drop=True), result.reset_index(drop=True)], axis=1)

# We will split the dataframe in smaller chunks 
# so we don't lose all progress if anything goes wrong
dfs_split = np.array_split(df, 100)

for index, chunk in enumerate(dfs_split):

  fname = f'gdrive/My Drive/DW - Climate Protests/9.high_engagement_language_detection/chunk-{index}.csv'

  if isfile(fname):
    continue

  start = time.time()
  
  result = lang_detect(chunk, classifier=classifier)
  result.to_csv(fname)

  end = time.time()
  print(f'Chunk {index} processing time:', str(timedelta(seconds=end-start)))

runtime.unassign()
