==> code/13.export_sentiment_analysis_dfs.ipynb <==
# Now we can take the results of the classification process and save then as csvs.

import glob
from google.colab import drive, runtime 
import pandas as pd

# Mounts the Google Drive so we can access the CSV files
drive.mount('/content/gdrive')

for index, folder in enumerate(["10.high_engagement_sentiment_analysis", "12.control_sentiment_analysis"]):

  # Reads all the CSV files
  fpaths = glob.glob(f"gdrive/My Drive/DW - Climate Protests/{folder}/*.csv")

  # Reads each csv file and concatenates
  df = pd.concat([ pd.read_csv(fpath) for fpath in fpaths]).reset_index(drop=True)

  if index == 0:
    outpath = "13.sentiment_analysis_dfs/protests.csv"
  elif index == 1:
    outpath = "13.sentiment_analysis_dfs/control.csv"

  df.to_csv(f"gdrive/My Drive/DW - Climate Protests/{outpath}")
